requirements.txt
flask==3.1.1
elasticsearch==9.0.1
python-dotenv==1.0.1
pandas>=1.5.0
requests>=2.28.0 